+++ zlasdtest/container/container.hpp
@@ -10,7 +10,7 @@
 
 void Empty(unsigned int&, unsigned int&, const lasd::Container&, bool);
 
-void Size(unsigned int&, unsigned int&, const lasd::Container&, bool, unsigned long);
+void Size(unsigned int&, unsigned int&, const lasd::Container&, bool, unsigned int);
 
 /* ************************************************************************** */
